refactoring directories for bitstring distribution and distance measures
