+++ src/python/zquantum/core/bitstring_distribution/__init__.py
@@ -1,0 +1,3 @@
+from ._bitstring_distribution import *
+from ._bitstring_distribution_test import *
+from .distance_measures import *
