+++ src/python/zquantum/core/bitstring_distribution/distance_measures/__init__.py
@@ -1,0 +1,2 @@
+from .clipped_negative_log_likelihood import *
+from .mmd import *
